+++ visualization/dashboard/src/MLIDashboard/Interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './IExplanationDashboardProps';
 export * from './IStringsParam';
+export * from './ITelemetryMessage'
